+++ src/scalar/index.ts
@@ -1,7 +1,11 @@
 import scalarElysiaTheme from './theme'
 import type { ReferenceConfiguration } from './types'
 
-export const ScalarRender = (version: string, config: ReferenceConfiguration, cdn: string) => `<!doctype html>
+export const ScalarRender = (
+    version: string,
+    config: ReferenceConfiguration,
+    cdn: string
+) => `<!doctype html>
 <html>
   <head>
     <title>API Reference</title>
@@ -21,7 +25,14 @@ export const ScalarRender = (version: string, config: ReferenceConfiguration, cd
   <body>
     <script
       id="api-reference"
-      data-configuration="JSON.stringify(${Bun.inspect(config)})"></script>
-    <script src="${cdn ? cdn:`https://cdn.jsdelivr.net/npm/@scalar/api-reference@${version}/dist/browser/standalone.min.js`}"></script>
+      data-url="${config.spec?.url}"
+      data-configuration='${JSON.stringify(config)}'
+    >
+    </script>
+    <script src="${
+        cdn
+            ? cdn
+            : `https://cdn.jsdelivr.net/npm/@scalar/api-reference@${version}/dist/browser/standalone.min.js`
+    }"></script>
   </body>
 </html>`